[ADD] GameManager
MainSceneManager 삭제
MainSceneManager 내용을GameManager에 병합
