--- Assets/Scripts/MainSceneManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60d103841495cde48b28c8dd4def3746
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
